Fix following splint warning.
src/cgen.c: (in function llvm_int8)
src/cgen.c:85:40: Function LLVMConstInt expects arg 2 to be unsigned
long long
                     gets int32_t: i
  To allow arbitrary integral types to match long unsigned, use
  +longunsignedintegral.

+++ src/cgen.c
@@ -80,7 +80,7 @@ static LLVMValueRef llvm_int1(bool b)
    return LLVMConstInt(LLVMInt1Type(), b, false);
 }
 
-static LLVMValueRef llvm_int8(int32_t i)
+static LLVMValueRef llvm_int8(int8_t i)
 {
    return LLVMConstInt(LLVMInt8Type(), i, false);
 }
